updated fork with machine name and commits

+++ case_copy.py
@@ -63,6 +63,7 @@ if (new_dir == orig_dir):
 #copy files to new directory
 os.system('mkdir -p '+new_dir+'/timing/checkpoints')
 os.system('rm  '+new_dir+'/*.nc')
+os.system('rm -f '+new_dir+'/*.nc.orig')
 os.system('cp  '+orig_dir+'/*_in* '+new_dir)
 os.system('cp  '+orig_dir+'/*nml '+new_dir)
 if (not ('ICB' in casename)):
@@ -128,11 +129,11 @@ for f in os.listdir(new_dir):
         myinput.close()
         os.system(' mv '+new_dir+'/'+f+'.tmp '+new_dir+'/'+f)
 #Assume makepointdata has been run to generate surface and domain data
-if (os.path.exists(orig_dir+'/surfdata.nc')):
+if (options.site_orig == options.site_new and os.path.exists(orig_dir+'/surfdata.nc')):
   os.system('cp '+orig_dir+'/surfdata.nc '+new_dir)
 else:
   os.system('cp temp/surfdata.nc '+new_dir)
-if (os.path.exists(orig_dir+'/domain.nc')):
+if (options.site_orig == options.site_new and os.path.exists(orig_dir+'/domain.nc')):
   os.system('cp '+orig_dir+'/domain.nc '+new_dir)
 else:
   os.system('cp temp/domain.nc '+new_dir)
